fix recursive overlay mounts
previously we were doing a single overlay mount, and simply having
/etc/fstab mount all non root mounts to /media/root-ro.  This had worked
in my cloud image testing, but failed to work on my laptop tests.  I
suspect that this was due to race conditions.

The issue was that when /home was on a separate partition, and that
mount got done over the top of the /media/root-ro/home directory, the initial
overlayfs mount did not then track the new /home mount (which actually makes
sense).

So now, we write out /etc/fstab so that each real mount point gets its own
overlayfs mount, that is done after the real mount is made to /media/root-ro.

This also adds 'nobootwait' to the overlaymount if the realmount already had it.

bzr-revno: 37