mount root as read-only if cmdline has 'ro'
The normal expected path of linux boot is for initramfs to lave the root
mounted as read-only.  That is specified with 'ro' in the kernel cmdline.

In order to make this look more like traditional boot, and potentially
avoid some bugs along the way (LP: #1031065) we also mount root read-only.

bzr-revno: 57