address LP: #1046405 explicitly
Newer maas cannot utilize 'ip=::::<hostname>:BOOTIF' functionality
that is provided by this package because it must operate with the
already released images that do not contain this package/function.

The workaround here is to allow newer maas to specify:
 ip=dhcp hostname=FOO BOOTIF=BAR
which will get translated through this package to:
 ip=::::FOO:NIC_NAME

Older images will continue to have it mean
 ip=:::::NIC_NAME

but will not break with multiple nics.

bzr-revno: 71