fix overlayroot and upstart
If upstart is being used, then you need to make sure that the overlay
path has an /etc/init.  Otherwise upstarts monitoring of /etc/init
will monitor the underlay directory and not get events to it.  The
end result of that is upstart not knowing about file creation or removal.

We check for presense of initctl in /sbin in the target, and allow it
to be a dangling link.  Also, we assume that if /etc/init does not exist
than this is not really upstart system.

if it is determined to be a upstart system, *always* create the marker file
that ensures that this target directory has the file (rather than just the
underlay having it).

LP: #1213925
bzr-revno: 87.1.1