overlayroot: fix overlayroot if recurse=0 was not provided.
When workdir support was added for kernels that needed it, it was only
added to the root filesystem mount.  Thus if there were more filesystems
and recurse=0 was not provided, then mounts would fail.

The fix here is to add the workdir entries and create directories for
all entries in the new /etc/fstab.

Also:
 * add and use 'clean_path' to make removing extraneous / from a path easier.
 * add and use needs_workdir function to indicate that workdir is needed
   on overlayfs mounts.
 * skip filesystems that are unsupported.  overlayroot cannot be backed
   by a vfat filesystem.   This would occur when trying to
   use it on /boot/efi in a cloud image.
 * use $MYTAG  in some places where string 'overlayroot' was used.

bzr-revno: 113