overlayroot: support random seed from systemd systemd-random-seed.service
The file /var/lib/urandom/random-seed is not being written in cloud-images
of ubuntu at 16.04.  Instead, the systemd provided service 
systemd-random-seed.service writes a random seed to
/var/lib/systemd/random-seed .

The change here just adjusts to use either file, preferring the
systemd one if both are there.

One other change is to provide a better debug message if read
failed on the seed file.

bzr-revno: 127