overlayroot: write the password to consistent filename
Previously, when overlayroot=crypt was used, and no password was
provided, the password was stored to a filename in /run/initramfs/
named overlayroot.XXXXXX.  The XXXXXX template was random.

This just made it more difficult to read that password file.

Now, we publish the passfile as /run/initramfs/overlayroot.password

LP: #1641678
bzr-revno: 129.1.1