overlayroot: Comment out the read-only line for /
The typical cloud image boot would have a line like:
  LABEL=cloudimg-rootfs / ext4 ro,defautls,noauto 0 0
and that would render 2 lines in the updated /etc/fstab:
  LABEL=cloudimg-rootfs /media/root-ro/ ext4 ro,defaults,noauto 0 0
  /media/root-ro/ / overlay lowerdir=...

The change here is to comment out the first line above.
This is to avoid systemd getting confused that /media/root-ro is
not required by / (since it is already mounted under /).

Also a removal of unused variable 'rel_file'.

LP: #1723183